workshop/agents-on-sagemaker/lab2/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anthropic==0.62.0
    # via langchain-anthropic
anyio==4.10.0
    # via
    #   anthropic
    #   httpx
    #   starlette
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
autopep8==2.3.2
    # via bedrock-agentcore-starter-toolkit
#bedrock-agentcore==0.1.1
    # via
    #   deepagents-agentcore (pyproject.toml)
    #   bedrock-agentcore-starter-toolkit
#bedrock-agentcore-starter-toolkit==0.1.5
    # via deepagents-agentcore (pyproject.toml)
boto3==1.40.5
    # via
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
    #   langchain-aws
botocore==1.40.5
    # via
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
    #   boto3
    #   s3transfer
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via prance
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   typer
    #   uvicorn
deepagents==0.0.3
    # via deepagents-agentcore (pyproject.toml)
distro==1.9.0
    # via anthropic
docstring-parser==0.17.0
    # via bedrock-agentcore-starter-toolkit
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   bedrock-agentcore-starter-toolkit
    #   langgraph-sdk
    #   langsmith
    #   tavily-python
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via bedrock-agentcore-starter-toolkit
jiter==0.10.0
    # via anthropic
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.0
    # via
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.4
    # via openapi-spec-validator
jsonschema-specifications==2025.4.1
    # via
    #   jsonschema
    #   openapi-schema-validator
langchain==0.3.27
    # via deepagents
langchain-anthropic==0.3.18
    # via deepagents
langchain-aws==0.2.30
    # via deepagents-agentcore (pyproject.toml)
langchain-core==0.3.74
    # via
    #   langchain
    #   langchain-anthropic
    #   langchain-aws
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-text-splitters==0.3.9
    # via langchain
langgraph==0.6.4
    # via deepagents
langgraph-checkpoint==2.1.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.6.4
    # via langgraph
langgraph-sdk==0.2.0
    # via langgraph
langsmith==0.4.13
    # via
    #   langchain
    #   langchain-core
lazy-object-proxy==1.11.0
    # via openapi-spec-validator
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==2.3.2
    # via langchain-aws
openapi-schema-validator==0.6.3
    # via openapi-spec-validator
openapi-spec-validator==0.7.2
    # via bedrock-agentcore-starter-toolkit
orjson==3.11.1
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.10.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   prance
pathable==0.4.4
    # via jsonschema-path
prance==25.4.8.0
    # via bedrock-agentcore-starter-toolkit
prompt-toolkit==3.0.51
    # via
    #   bedrock-agentcore-starter-toolkit
    #   questionary
py-openapi-schema-to-json-schema==0.0.3
    # via bedrock-agentcore-starter-toolkit
pycodestyle==2.14.0
    # via autopep8
pydantic==2.11.7
    # via
    #   anthropic
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
    #   langchain
    #   langchain-anthropic
    #   langchain-aws
    #   langchain-core
    #   langgraph
    #   langsmith
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via botocore
pyyaml==6.0.2
    # via
    #   bedrock-agentcore-starter-toolkit
    #   jsonschema-path
    #   langchain
    #   langchain-core
questionary==2.1.0
    # via bedrock-agentcore-starter-toolkit
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
regex==2025.7.34
    # via tiktoken
requests==2.32.4
    # via
    #   bedrock-agentcore-starter-toolkit
    #   jsonschema-path
    #   langchain
    #   langsmith
    #   prance
    #   requests-toolbelt
    #   tavily-python
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rich==14.1.0
    # via
    #   bedrock-agentcore-starter-toolkit
    #   typer
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.14
    # via
    #   bedrock-agentcore-starter-toolkit
    #   prance
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
s3transfer==0.13.1
    # via boto3
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
sqlalchemy==2.0.42
    # via langchain
starlette==0.47.2
    # via bedrock-agentcore
tavily-python==0.7.10
    # via deepagents-agentcore (pyproject.toml)
tenacity==9.1.2
    # via langchain-core
tiktoken==0.10.0
    # via tavily-python
toml==0.10.2
    # via bedrock-agentcore-starter-toolkit
typer==0.16.0
    # via bedrock-agentcore-starter-toolkit
typing-extensions==4.14.1
    # via
    #   anthropic
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
    #   botocore
    #   requests
uvicorn==0.35.0
    # via
    #   bedrock-agentcore
    #   bedrock-agentcore-starter-toolkit
wcwidth==0.2.13
    # via prompt-toolkit
xxhash==3.5.0
    # via langgraph
zstandard==0.23.0
    # via langsmith
bedrock-agentcore
bedrock-agentcore-starter-toolkit==0.1.1
uv
langchain[aws]
langgraph
langsmith[otel]
duckduckgo-search 
langchain-community
opentelemetry-instrumentation-langchain
starlette
uvicorn
boto3
botocore
yfinance
pandas
numpy
requests
reportlab
matplotlib
seaborn
python-dateutil
pytz
